requirements.txt
paddlepaddle
paddleocr
pix2tex[gui]
streamlit
opencv-python-headless
Pillow
pdf2image
numpy
